+++ core/src/main/java/com/graphhopper/routing/util/parsers/BikeCommonAverageSpeedParser.java
@@ -1,6 +1,7 @@
 package com.graphhopper.routing.util.parsers;
 
 import com.graphhopper.reader.ReaderWay;
+import com.graphhopper.routing.ev.BooleanEncodedValue;
 import com.graphhopper.routing.ev.DecimalEncodedValue;
 import com.graphhopper.routing.ev.EnumEncodedValue;
 import com.graphhopper.routing.ev.EdgeIntAccess;
@@ -155,8 +156,8 @@ public abstract class BikeCommonAverageSpeedParser extends AbstractAverageSpeedP
         double speedFwd = applyMaxSpeed(way, speed, false);
         avgSpeedEnc.setDecimal(false, edgeId, edgeIntAccess, speedFwd);
         if (avgSpeedEnc.isStoreTwoDirections()) {
-            speed = applyMaxSpeed(way, speed, true);
-            avgSpeedEnc.setDecimal(true, edgeId, edgeIntAccess, speed);
+            double bwdSpeed = applyMaxSpeed(way, speed, true);
+            avgSpeedEnc.setDecimal(true, edgeId, edgeIntAccess, bwdSpeed);
         }
     }
 
@@ -167,7 +168,7 @@ public abstract class BikeCommonAverageSpeedParser extends AbstractAverageSpeedP
 
         if (way.hasTag("railway", "platform"))
             highwaySpeed = PUSHING_SECTION_SPEED;
-        // Under certain conditions we need to increase the speed of pushing sections to the speed of a "highway=cycleway"
+            // Under certain conditions we need to increase the speed of pushing sections to the speed of a "highway=cycleway"
         else if (way.hasTag("highway", pushingSectionsHighways)
                 && ((way.hasTag("foot", "yes") && way.hasTag("segregated", "yes"))
                 || (way.hasTag("bicycle", intendedValues))))
@@ -180,12 +181,8 @@ public abstract class BikeCommonAverageSpeedParser extends AbstractAverageSpeedP
             if (surfaceSpeed != null) {
                 speed = surfaceSpeed;
                 // boost handling for good surfaces but avoid boosting if pushing section
-                if (highwaySpeed != null && surfaceSpeed > highwaySpeed) {
-                    if (pushingSectionsHighways.contains(highwayTag))
-                        speed = highwaySpeed;
-                    else
-                        speed = surfaceSpeed;
-                }
+                if (highwaySpeed != null && surfaceSpeed > highwaySpeed && pushingSectionsHighways.contains(highwayTag))
+                    speed = highwaySpeed;
             }
         } else {
             String tt = way.getTag("tracktype");