+++ core/src/main/java/com/graphhopper/routing/util/parsers/OSMGetOffBikeParser.java
@@ -10,29 +10,39 @@ import java.util.HashSet;
 import java.util.List;
 
 /**
- * This parser scans different OSM tags to identify ways where a cyclist has to get off her bike.
+ * This parser scans different OSM tags to identify ways where a cyclist has to get off her bike. Like on footway but
+ * also in reverse oneway direction.
  */
 public class OSMGetOffBikeParser implements TagParser {
-    private final HashSet<String> pushBikeHighwayTags = new HashSet<>();
-    private final List<String> accepted = Arrays.asList("designated", "yes", "official", "permissive");
-    private final BooleanEncodedValue offBikeEnc;
 
-    public OSMGetOffBikeParser(BooleanEncodedValue getOffBikeEnc) {
-        // steps -> special handling
-        this(getOffBikeEnc, Arrays.asList("path", "footway", "pedestrian", "platform"));
-    }
+    private final List<String> INTENDED = Arrays.asList("designated", "yes", "official", "permissive");
+    // steps -> special handling
+    private final HashSet<String> GET_OFF_BIKE = new HashSet<>(Arrays.asList("path", "footway", "pedestrian", "platform"));
+    private final BooleanEncodedValue getOffBikeEnc;
+    private final BooleanEncodedValue bikeAccessEnc;
 
-    public OSMGetOffBikeParser(BooleanEncodedValue getOffBikeEnc, List<String> pushBikeTags) {
-        offBikeEnc = getOffBikeEnc;
-        pushBikeHighwayTags.addAll(pushBikeTags);
+    /**
+     * @param bikeAccessEnc used to find out if way is oneway and so it does not matter which bike type is used.
+     */
+    public OSMGetOffBikeParser(BooleanEncodedValue getOffBikeEnc, BooleanEncodedValue bikeAccessEnc) {
+        this.getOffBikeEnc = getOffBikeEnc;
+        this.bikeAccessEnc = bikeAccessEnc;
     }
 
     @Override
     public void handleWayTags(int edgeId, EdgeIntAccess edgeIntAccess, ReaderWay way, IntsRef relationFlags) {
         String highway = way.getTag("highway");
-        if (!way.hasTag("bicycle", accepted) && (pushBikeHighwayTags.contains(highway) || way.hasTag("railway", "platform"))
+        if (!way.hasTag("bicycle", INTENDED) && (GET_OFF_BIKE.contains(highway) || way.hasTag("railway", "platform"))
                 || "steps".equals(highway) || way.hasTag("bicycle", "dismount")) {
-            offBikeEnc.setBool(false, edgeId, edgeIntAccess, true);
+            getOffBikeEnc.setBool(false, edgeId, edgeIntAccess, true);
+            getOffBikeEnc.setBool(true, edgeId, edgeIntAccess, true);
+        }
+        boolean fwd = bikeAccessEnc.getBool(false, edgeId, edgeIntAccess);
+        boolean bwd = bikeAccessEnc.getBool(true, edgeId, edgeIntAccess);
+        // get off bike for reverse oneways
+        if (fwd != bwd) {
+            if (!fwd) getOffBikeEnc.setBool(false, edgeId, edgeIntAccess, true);
+            if (!bwd) getOffBikeEnc.setBool(true, edgeId, edgeIntAccess, true);
         }
     }
 }