+++ core/src/test/java/com/graphhopper/routing/util/parsers/OSMGetOffBikeParserTest.java
@@ -1,8 +1,15 @@
 package com.graphhopper.routing.util.parsers;
 
 import com.graphhopper.reader.ReaderWay;
-import com.graphhopper.routing.ev.*;
+import com.graphhopper.reader.osm.conditional.DateRangeParser;
+import com.graphhopper.routing.ev.ArrayEdgeIntAccess;
+import com.graphhopper.routing.ev.BooleanEncodedValue;
+import com.graphhopper.routing.ev.EdgeIntAccess;
+import com.graphhopper.routing.ev.GetOffBike;
+import com.graphhopper.routing.util.EncodingManager;
+import com.graphhopper.routing.util.VehicleEncodedValues;
 import com.graphhopper.storage.IntsRef;
+import com.graphhopper.util.PMap;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -10,10 +17,14 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 public class OSMGetOffBikeParserTest {
     private final BooleanEncodedValue offBikeEnc = GetOffBike.create();
-    private final OSMGetOffBikeParser parser = new OSMGetOffBikeParser(offBikeEnc);
+    private final BikeAccessParser accessParser;
+    private final OSMGetOffBikeParser getOffParser;
 
     public OSMGetOffBikeParserTest() {
-        offBikeEnc.init(new EncodedValue.InitializerConfig());
+        EncodingManager em = new EncodingManager.Builder().add(offBikeEnc).add(VehicleEncodedValues.bike(new PMap()).getAccessEnc()).build();
+        accessParser = new BikeAccessParser(em, new PMap());
+        accessParser.init(new DateRangeParser());
+        getOffParser = new OSMGetOffBikeParser(offBikeEnc, accessParser.getAccessEnc());
     }
 
     @Test
@@ -96,11 +107,28 @@ public class OSMGetOffBikeParserTest {
         assertTrue(isGetOffBike(way));
     }
 
+    @Test
+    public void testOneway() {
+        ReaderWay way = new ReaderWay(1);
+        way.setTag("highway", "primary");
+        way.setTag("oneway", "yes");
+
+        EdgeIntAccess edgeIntAccess = new ArrayEdgeIntAccess(1);
+        int edgeId = 0;
+        IntsRef rel = new IntsRef(1);
+        accessParser.handleWayTags(edgeId, edgeIntAccess, way, new IntsRef(1));
+        getOffParser.handleWayTags(edgeId, edgeIntAccess, way, new IntsRef(1));
+
+        assertFalse(offBikeEnc.getBool(false, edgeId, edgeIntAccess));
+        assertTrue(offBikeEnc.getBool(true, edgeId, edgeIntAccess));
+    }
+
     private boolean isGetOffBike(ReaderWay way) {
         EdgeIntAccess edgeIntAccess = new ArrayEdgeIntAccess(1);
         int edgeId = 0;
-        IntsRef relationFlags = new IntsRef(1);
-        parser.handleWayTags(edgeId, edgeIntAccess, way, relationFlags);
+        IntsRef rel = new IntsRef(1);
+        accessParser.handleWayTags(edgeId, edgeIntAccess, way, rel);
+        getOffParser.handleWayTags(edgeId, edgeIntAccess, way, rel);
         return offBikeEnc.getBool(false, edgeId, edgeIntAccess);
     }
 }