+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### 8.0 [not yet released]
+
 - renamed EdgeKVStorage to KVStorage as it is (temporarily) used for node tage too, see #2705
+- bike vehicles are now allowed to go in reverse direction of oneways, see custom_models/bike.json #196
 
 ### 7.0 [14 Mar 2023]